% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{FIREDAM data set.}
\format{
A data frame with 15 rows and 2 variables:
\describe{
  \item{Distance}
  \item{Damage}
  ...
}
}
\source{
\url{https://www.crcpress.com/Statistics-for-Engineering-and-the-Sciences/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
fire
}
\description{
A data set collecting the number of fires and their damage
}
\keyword{datasets}
