% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATH4753Hart0009.R
\docType{package}
\name{MATH4753Hart0009}
\alias{MATH4753Hart0009}
\title{BMMI Package}
\description{
Contains one function for creating BMI groups.
}
\author{
Hanna Hart \email{hannaahart5@gmail.com}
}
