% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{mybin}
\usage{
mybin(iter = 1000, n = 10, p = 0.5)
}
\arguments{
\item{p}{}
}
\value{
barplot based on parameters passed in
}
\description{
mybin
}
\examples{
mybin(iter = 1000, n = 50, p =0.2)
}
