% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{myci}
\usage{
myci(iter = 10000, x1, x2, fun = "mean", alpha = 0.05, ...)
}
\arguments{
\item{...}{}
}
\value{
returns a histogram with the confidence interval of the samples the user passes through
}
\description{
myci
}
\examples{
 myci(x1 = c(1,2,3,4), x2 = c(5,6,7,8))
}
