% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{myctlp}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{...}{}
}
\value{
returns histogram of the density function, barplot of relative frequency, and a barplot of probability
}
\description{
myctlp
}
\examples{
myctlp(n = 10, iter = 100000, lambda = 10)
}
